# server/public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GroChain Backend API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for GroChain backend."
  },
  "servers": [
    { "url": "http://localhost:5000" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Harvest": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "farmer": { "type": "string" },
          "cropType": { "type": "string" },
          "quantity": { "type": "number" },
          "unit": { "type": "string", "enum": ["kg","tons","pieces","bundles","bags","crates"] },
          "date": { "type": "string", "format": "date-time" },
          "geoLocation": { "type": "object", "properties": { "lat": {"type":"number"}, "lng": {"type":"number"} } },
          "batchId": { "type": "string" },
          "qrData": { "type": "string" },
          "quality": { "type": "string", "enum": ["excellent","good","fair","poor"] }
        }
      },
...
    }
  },
  "security": [ { "bearerAuth": [] } ],
  "paths": {
...
    "/api/auth/register": { "post": { "summary": "Register user", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "name": {"type":"string"}, "email": {"type":"string"}, "phone": {"type":"string"}, "password": {"type":"string"}, "role": {"type":"string"} }, "required": ["name","email","phone","password"] } } } }, "responses": { "201": { "description": "Created" } } } },
    "/api/auth/login": { "post": { "summary": "Login", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": {"type":"string"}, "password": {"type":"string"} }, "required": ["email","password"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/refresh": { "post": { "summary": "Refresh tokens", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "refreshToken": {"type":"string"} }, "required": ["refreshToken"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/forgot-password": { "post": { "summary": "Forgot password", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "email": {"type":"string"} }, "required": ["email"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/reset-password": { "post": { "summary": "Reset password", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "token": {"type":"string"}, "password": {"type":"string"} }, "required": ["token","password"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/verify-email": { "post": { "summary": "Verify email", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "token": {"type":"string"} }, "required": ["token"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/send-sms-otp": { "post": { "summary": "Send SMS OTP", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "phone": {"type":"string"} }, "required": ["phone"] } } } }, "responses": { "200": { "description": "OK" } } } },
    "/api/auth/verify-sms-otp": { "post": { "summary": "Verify SMS OTP", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "properties": { "phone": {"type":"string"}, "otp": {"type":"string"} }, "required": ["phone","otp"] } } } }, "responses": { "200": { "description": "OK" } } } },
...
    "/api/harvests": { "get": { "summary": "List my harvests", "parameters": [ {"name":"cropType","in":"query","schema":{"type":"string"}}, {"name":"page","in":"query","schema":{"type":"integer"}}, {"name":"limit","in":"query","schema":{"type":"integer"}} ], "responses": { "200": { "description": "OK" } } }, "post": { "summary": "Create harvest", "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Harvest" } } } }, "responses": { "201": { "description": "Created" } } } },
...
  }
}